+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/composer/MessageComposer.kt
@@ -48,8 +48,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.Bottom
 import androidx.compose.ui.Alignment.Companion.Center
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLayoutDirection
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
@@ -171,7 +174,16 @@ public fun MessageComposer(
     },
     input: @Composable RowScope.(MessageComposerState) -> Unit = {
         with(ChatTheme.componentFactory) {
+            val focusRequester = remember { FocusRequester() }
+            val keyboardController = LocalSoftwareKeyboardController.current
+            LaunchedEffect(Unit) {
+                viewModel.focusRequestFlow.collect {
+                    focusRequester.requestFocus()
+                    keyboardController?.show()
+                }
+            }
             MessageComposerInput(
+                modifier = Modifier.focusRequester(focusRequester),
                 state = it,
                 onInputChanged = onValueChange,
                 onAttachmentRemoved = onAttachmentRemoved,
@@ -319,6 +331,7 @@ public fun MessageComposer(
     input: @Composable RowScope.(MessageComposerState) -> Unit = {
         with(ChatTheme.componentFactory) {
             MessageComposerInput(
+                modifier = Modifier,
                 state = it,
                 onInputChanged = onValueChange,
                 onAttachmentRemoved = onAttachmentRemoved,